Make build configuration more robust.
Found a trick that allows us to detect if the Qt we're building
against is shared or static.  In the static case, we want to link
in the image plugins... but this is a bad thing for the shared
case.
Also started using a mac-specific conditional for a check that
only makes sense on macs anyway.


git-svn-id: https://gforge.sci.utah.edu/svn/imagevis3d/CmdLineConverter@1030 c36c8488-0289-0348-9b64-b301f74bd9a7