Make the command line converter a bundle on OS X.
This is annoying, because one needs to invoke a command line app by
manually navigating into the bundle.  If it's not a bundle, though,
we can't access Qt on the user's system.

Long term, it would be nice to completely remove Qt from Tuvok for
other reasons, and it would solve this problem as well.  We're a
long way from that ideal, though...


git-svn-id: https://gforge.sci.utah.edu/svn/imagevis3d/CmdLineConverter@1468 c36c8488-0289-0348-9b64-b301f74bd9a7